Added external text editor support
The configuration file now includes an 'abort_text' and
'text_edit_command' option. The latter is the command to run to start up
the preferred text editor. It will edit a temporary file, so you must
save the file before exiting. This is intended to be utilized with a
command that halts the console, thus the program will not resume until
the editor closes.

Since the editor is no longer in console, the program cannot be killed
with console commands. Therefore the 'abort_text' option, when saved by
the editor, will raise an Exception to exit the program.